chore: Move swapper functions into useSwapper
